requirements.txt
discord.py>=2.0.0
croniter>=6.0.0
python-dotenv

